:truck: Split longitudinal into anat and func \& :rotating_light: Lint anat
