+++ CPAC/longitudinal/wf/__init__.py
@@ -1,0 +1,25 @@
+# Copyright (C) 2024  C-PAC Developers
+
+# This file is part of C-PAC.
+
+# C-PAC is free software: you can redistribute it and/or modify it under
+# the terms of the GNU Lesser General Public License as published by the
+# Free Software Foundation, either version 3 of the License, or (at your
+# option) any later version.
+
+# C-PAC is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+# FITNESS FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public
+# License for more details.
+
+# You should have received a copy of the GNU Lesser General Public
+# License along with C-PAC. If not, see <https://www.gnu.org/licenses/>.
+"""Workflows for longitudinal preprocessing."""
+
+from CPAC.utils.docs import DOCS_URL_PREFIX
+
+assert isinstance(__doc__, str)
+__doc__ += f"""
+
+See {DOCS_URL_PREFIX}/user/longitudinal
+"""  # noqa: A001
